for #660: refactor AggregationSelectItemAssert aggregationType to type
