+++ sharding-core/src/test/resources/parser/select_aggregate.xml
@@ -8,7 +8,7 @@
             <table-token begin-position="40" original-literals="t_order" />
         </tokens>
         <aggregation-select-items>
-            <aggregation-select-item inner-expression="(user_id)" aggregation-type="SUM" alias="user_id_sum" />
+            <aggregation-select-item type="SUM" inner-expression="(user_id)" alias="user_id_sum" />
         </aggregation-select-items>
     </parser-assert>
     
@@ -20,7 +20,7 @@
             <table-token begin-position="37" original-literals="t_order" />
         </tokens>
         <aggregation-select-items>
-            <aggregation-select-item inner-expression="(*)" aggregation-type="COUNT" alias="orders_count" />
+            <aggregation-select-item type="COUNT" inner-expression="(*)" alias="orders_count" />
         </aggregation-select-items>
     </parser-assert>
     
@@ -32,7 +32,7 @@
             <table-token begin-position="40" original-literals="t_order" />
         </tokens>
         <aggregation-select-items>
-            <aggregation-select-item inner-expression="(user_id)" aggregation-type="MAX" alias="max_user_id" />
+            <aggregation-select-item type="MAX" inner-expression="(user_id)" alias="max_user_id" />
         </aggregation-select-items>
     </parser-assert>
     
@@ -44,7 +44,7 @@
             <table-token begin-position="40" original-literals="t_order" />
         </tokens>
         <aggregation-select-items>
-            <aggregation-select-item inner-expression="(user_id)" aggregation-type="MIN" alias="min_user_id" />
+            <aggregation-select-item type="MIN" inner-expression="(user_id)" alias="min_user_id" />
         </aggregation-select-items>
     </parser-assert>
     
@@ -62,9 +62,9 @@
             </items-token>
         </tokens>
         <aggregation-select-items>
-            <aggregation-select-item inner-expression="(user_id)" aggregation-type="AVG" alias="user_id_avg">
-                <derived-column inner-expression="(user_id)" aggregation-type="COUNT" alias="AVG_DERIVED_COUNT_0" />
-                <derived-column inner-expression="(user_id)" aggregation-type="SUM" alias="AVG_DERIVED_SUM_0" />
+            <aggregation-select-item type="AVG" inner-expression="(user_id)" alias="user_id_avg">
+                <derived-column type="COUNT" inner-expression="(user_id)" alias="AVG_DERIVED_COUNT_0" />
+                <derived-column type="SUM" inner-expression="(user_id)" alias="AVG_DERIVED_SUM_0" />
             </aggregation-select-item>
         </aggregation-select-items>
     </parser-assert>
@@ -77,7 +77,7 @@
             <table-token begin-position="46" original-literals="t_order" />
         </tokens>
         <aggregation-select-items>
-            <aggregation-select-item inner-expression="(`order_id`)" aggregation-type="COUNT" alias="orders_count" />
+            <aggregation-select-item type="COUNT" inner-expression="(`order_id`)" alias="orders_count" />
         </aggregation-select-items>
     </parser-assert>
 </parser-asserts>
