+++ sharding-core/src/test/resources/parser/select_group_by.xml
@@ -15,7 +15,7 @@
             </condition>
         </conditions>
         <aggregation-select-items>
-            <aggregation-select-item inner-expression="(*)" aggregation-type="COUNT" alias="c_number" />
+            <aggregation-select-item type="COUNT" inner-expression="(*)" alias="c_number" />
         </aggregation-select-items>
         <group-by-columns>
             <group-by-column name="date_format(c_date, '%y-%m-%d')" order-direction="ASC" alias="c_date" />
@@ -33,7 +33,7 @@
             <table-token begin-position="49" original-literals="t_order" />
         </tokens>
         <aggregation-select-items>
-            <aggregation-select-item inner-expression="(order_id)" aggregation-type="SUM" alias="orders_sum" />
+            <aggregation-select-item type="SUM" inner-expression="(order_id)" alias="orders_sum" />
         </aggregation-select-items>
         <group-by-columns>
             <group-by-column name="user_id" order-direction="ASC" />
@@ -51,7 +51,7 @@
             <table-token begin-position="53" original-literals="t_order" />
         </tokens>
         <aggregation-select-items>
-            <aggregation-select-item inner-expression="(order_id)" aggregation-type="COUNT" alias="orders_count" />
+            <aggregation-select-item type="COUNT" inner-expression="(order_id)" alias="orders_count" />
         </aggregation-select-items>
         <group-by-columns>
             <group-by-column name="user_id" order-direction="ASC" />
@@ -69,7 +69,7 @@
             <table-token begin-position="51" original-literals="t_order" />
         </tokens>
         <aggregation-select-items>
-            <aggregation-select-item inner-expression="(order_id)" aggregation-type="MAX" alias="max_order_id" />
+            <aggregation-select-item type="MAX" inner-expression="(order_id)" alias="max_order_id" />
         </aggregation-select-items>
         <group-by-columns>
             <group-by-column name="user_id" order-direction="ASC" />
@@ -87,7 +87,7 @@
             <table-token begin-position="51" original-literals="t_order" />
         </tokens>
         <aggregation-select-items>
-            <aggregation-select-item inner-expression="(order_id)" aggregation-type="MIN" alias="min_order_id"/>
+            <aggregation-select-item type="MIN" inner-expression="(order_id)" alias="min_order_id"/>
         </aggregation-select-items>
         <group-by-columns>
             <group-by-column name="user_id" order-direction="ASC" />
@@ -111,9 +111,9 @@
             </items-token>
         </tokens>
         <aggregation-select-items>
-            <aggregation-select-item inner-expression="(order_id)" aggregation-type="AVG" alias="orders_avg">
-                <derived-column inner-expression="(order_id)" aggregation-type="COUNT" alias="AVG_DERIVED_COUNT_0" />
-                <derived-column inner-expression="(order_id)" aggregation-type="SUM" alias="AVG_DERIVED_SUM_0" />
+            <aggregation-select-item type="AVG" inner-expression="(order_id)" alias="orders_avg">
+                <derived-column type="COUNT" inner-expression="(order_id)" alias="AVG_DERIVED_COUNT_0" />
+                <derived-column type="SUM" inner-expression="(order_id)" alias="AVG_DERIVED_SUM_0" />
             </aggregation-select-item>
         </aggregation-select-items>
         <group-by-columns>
@@ -132,7 +132,7 @@
             <table-token begin-position="49" original-literals="t_order" />
         </tokens>
         <aggregation-select-items>
-            <aggregation-select-item inner-expression="(order_id)" aggregation-type="SUM" alias="orders_sum" />
+            <aggregation-select-item type="SUM" inner-expression="(order_id)" alias="orders_sum" />
         </aggregation-select-items>
         <group-by-columns>
             <group-by-column name="user_id" order-direction="ASC" />
@@ -168,7 +168,7 @@
             </condition>
         </conditions>
         <aggregation-select-items>
-            <aggregation-select-item inner-expression="(*)" aggregation-type="COUNT" alias="items_count" />
+            <aggregation-select-item type="COUNT" inner-expression="(*)" alias="items_count" />
         </aggregation-select-items>
         <group-by-columns>
             <group-by-column name="user_id" alias="GROUP_BY_DERIVED_0" owner="o" order-direction="ASC" />
@@ -198,7 +198,7 @@
             </condition>
         </conditions>
         <aggregation-select-items>
-            <aggregation-select-item inner-expression="(*)" aggregation-type="COUNT" alias="items_count" />
+            <aggregation-select-item type="COUNT" inner-expression="(*)" alias="items_count" />
         </aggregation-select-items>
         <group-by-columns>
             <group-by-column name="user_id" owner="o" order-direction="ASC" />
@@ -234,7 +234,7 @@
             <row-count-token row-count="5" begin-position="89" />
         </tokens>
         <aggregation-select-items>
-            <aggregation-select-item inner-expression="(order_id)" aggregation-type="SUM"/>
+            <aggregation-select-item type="SUM" inner-expression="(order_id)" />
         </aggregation-select-items>
         <group-by-columns>
             <group-by-column name="user_id" order-direction="ASC" />
@@ -269,7 +269,7 @@
             <order-by-token begin-position-with-placeholder="80" begin-position-without-placeholder="80" />
         </tokens>
         <aggregation-select-items>
-            <aggregation-select-item inner-expression="(order_id)" aggregation-type="SUM" alias="orders_sum" />
+            <aggregation-select-item type="SUM" inner-expression="(order_id)" alias="orders_sum" />
         </aggregation-select-items>
         <group-by-columns>
             <group-by-column name="key" alias="key" order-direction="ASC" />
